Created service and dto packages and moved respective files into packages
The reasoning for this is just for easier organization and separation of different java files
