+++ src/main/java/com/timelogr/enterprise/dto/employee.java
@@ -1,4 +1,4 @@
-package com.timelogr.enterprise;
+package com.timelogr.enterprise.dto;
 
 public class employee {
     private long id;
