+++ src/main/java/com/timelogr/enterprise/dto/timeLog.java
@@ -1,4 +1,4 @@
-package com.timelogr.enterprise;
+package com.timelogr.enterprise.dto;
 
 public class timeLog {
     private long id;
